+++ app/api/auth/[...nextauth]/route.ts
@@ -44,9 +44,6 @@ export const authOptions:NextAuthOptions = {
       return session;
     },
 
-    async redirect(){
-      return '/auth/post-login';
-    }
   },
 
   pages: {
